requirements.txt
streamlit>=1.44
pandas>=2.2
python-dotenv>=1.0
openai>=1.30
pytest>=8.0
